refactor(core): Organize Redis under scaling mode
- Move `/redis` under scaling mode
- Move channel constants to `scaling/constants.ts`
- Delete `redis-service-base-classes.ts` - unused since #10616
- Delete `WorkerJobStatusSummary` - unused
- Move pubsub map to `pubsub.types.ts`
- Rename `PubSubMessage` to `JobReport` for consistency
