+++ packages/cli/src/scaling/constants.ts
@@ -1,3 +1,7 @@
 export const QUEUE_NAME = 'jobs';
 
 export const JOB_TYPE_NAME = 'job';
+
+export const COMMAND_PUBSUB_CHANNEL = 'n8n.commands';
+
+export const WORKER_RESPONSE_PUBSUB_CHANNEL = 'n8n.worker-response';
