+++ packages/cli/src/scaling/pubsub/subscriber.service.ts
@@ -3,7 +3,7 @@ import { Service } from 'typedi';
 
 import config from '@/config';
 import { Logger } from '@/logger';
-import { RedisClientService } from '@/services/redis/redis-client.service';
+import { RedisClientService } from '@/scaling/redis/redis-client.service';
 
 import type { ScalingPubSubChannel } from './pubsub.types';
