+++ packages/cli/src/scaling/redis/redis-service-commands.ts
