+++ packages/cli/src/services/__tests__/orchestration.service.test.ts
@@ -9,13 +9,13 @@ import config from '@/config';
 import { MessageEventBus } from '@/eventbus/message-event-bus/message-event-bus';
 import { ExternalSecretsManager } from '@/external-secrets/external-secrets-manager.ee';
 import { Push } from '@/push';
+import { RedisClientService } from '@/scaling/redis/redis-client.service';
+import type { RedisServiceWorkerResponseObject } from '@/scaling/redis/redis-service-commands';
 import * as helpers from '@/services/orchestration/helpers';
 import { handleCommandMessageMain } from '@/services/orchestration/main/handle-command-message-main';
 import { handleWorkerResponseMessageMain } from '@/services/orchestration/main/handle-worker-response-message-main';
 import { OrchestrationHandlerMainService } from '@/services/orchestration/main/orchestration.handler.main.service';
 import { OrchestrationService } from '@/services/orchestration.service';
-import { RedisClientService } from '@/services/redis/redis-client.service';
-import type { RedisServiceWorkerResponseObject } from '@/services/redis/redis-service-commands';
 import { mockInstance } from '@test/mocking';
 
 import type { MainResponseReceivedHandlerOptions } from '../orchestration/main/types';
