+++ packages/cli/src/services/orchestration/worker/types.ts
@@ -1,5 +1,4 @@
 import type { RunningJobSummary } from '@n8n/api-types';
-import type { ExecutionStatus, WorkflowExecuteMode } from 'n8n-workflow';
 
 import type { Publisher } from '@/scaling/pubsub/publisher.service';
 
@@ -8,15 +7,4 @@ export interface WorkerCommandReceivedHandlerOptions {
 	publisher: Publisher;
 	getRunningJobIds: () => Array<string | number>;
 	getRunningJobsSummary: () => RunningJobSummary[];
-}
-
-export interface WorkerJobStatusSummary {
-	jobId: string;
-	executionId: string;
-	retryOf?: string;
-	startedAt: Date;
-	mode: WorkflowExecuteMode;
-	workflowName: string;
-	workflowId: string;
-	status: ExecutionStatus;
 }
