--- packages/cli/src/services/redis/redis-constants.ts
@@ -1,2 +1,0 @@
-export const COMMAND_REDIS_CHANNEL = 'n8n.commands';
-export const WORKER_RESPONSE_REDIS_CHANNEL = 'n8n.worker-response';
